requirements.txt
# Bot framework
python-telegram-bot[job-queue]>=21.0

# Library for automating user accounts (Telethon client)
telethon==1.34.0

# Web server for webhook
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# For rich, colorful logging
rich==13.7.1

# Database ORM
SQLAlchemy==2.0.29